# docs/swagger.yaml
basePath: /
definitions:
...
  models.AsetSnapshot:
    properties:
      aset_bersih:
        type: number
      bulan:
        description: 'Contoh: 2026-05-01'
        type: string
      created_at:
        type: string
      id:
        type: integer
      total_hutang:
        type: number
      total_kas:
        type: number
      total_persediaan:
        type: number
      total_piutang:
        type: number
    type: object
  models.Bahan:
    properties:
      batas_minimum:
        type: number
      harga_saat_ini:
        description: Update otomatis dari pembelian terakhir
        type: number
      id:
        type: integer
      nama_bahan:
        type: string
      satuan:
        description: gr, ml, pcs
        type: string
      stok:
        type: number
      urutan:
        type: integer
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: Terjadi kesalahan pada server
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        example: Operasi berhasil dilakukan
        type: string
    type: object
...
  models.SnapshotAsetInput:
    properties:
      bulan:
        example: "2026-05-01"
        type: string
    required:
    - bulan
    type: object
...
  models.UrutanBahanInput:
    properties:
      id:
        example: 1
        type: integer
      urutan:
        example: 2
        type: integer
    required:
    - id
    - urutan
    type: object
...
host: localhost:3000
info:
  contact:
    name: Aristo Budiman
  description: Dokumentasi API Internal untuk Sistem Inventory, Nota, dan Kas.
  title: Tiara Bakery Master API
  version: "1.0"
paths:
...
  /api/aset/riwayat:
    get:
      consumes:
      - application/json
      description: Menarik riwayat snapshot kekayaan per bulan, diurutkan dari yang
        terbaru.
      produces:
      - application/json
      responses:
        "200":
          description: Data riwayat aset ditarik
          schema:
            items:
              $ref: '#/definitions/models.AsetSnapshot'
            type: array
      security:
      - BearerAuth: []
      summary: Riwayat Pertumbuhan Aset
      tags:
      - 15. Analisis & Aset
  /api/aset/snapshot:
    post:
      consumes:
      - application/json
      description: Mengunci dan menyimpan nilai kalkulasi aset akhir bulan ke dalam
        riwayat, agar pertumbuhan kekayaan (profit bersih bulanan) bisa diukur dengan
        akurat.
      parameters:
      - description: Tanggal / Bulan kunci
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.SnapshotAsetInput'
      produces:
      - application/json
      responses:
        "200":
          description: Snapshot berhasil dikunci
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Format tanggal salah
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kunci Snapshot Aset (Tutup Bulan)
      tags:
      - 15. Analisis & Aset
  /api/bahan:
    get:
      consumes:
      - application/json
      description: Mengambil daftar master bahan baku gudang, diurutkan berdasarkan
        urutan tampilan.
      produces:
      - application/json
      responses:
        "200":
          description: List data bahan berhasil ditarik
          schema:
            items:
              $ref: '#/definitions/models.Bahan'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil Seluruh Master Bahan
      tags:
      - 05. Master Bahan Baku
    post:
      consumes:
      - application/json
      description: Menyimpan data master bahan baku baru ke gudang.
      parameters:
      - description: Data Bahan utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BahanInput'
      produces:
      - application/json
      responses:
        "200":
          description: Bahan berhasil dibuat
          schema:
            $ref: '#/definitions/models.Bahan'
        "400":
          description: Format JSON tidak valid
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Kesalahan eksekusi database
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat Master Bahan Baru
      tags:
      - 05. Master Bahan Baku
  /api/bahan/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus master bahan baku dari daftar aktif.
      parameters:
      - description: ID Bahan
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Bahan berhasil dihapus
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus Master Bahan (Soft Delete)
      tags:
      - 05. Master Bahan Baku
    put:
      consumes:
      - application/json
      description: Memperbarui profil, satuan, atau harga bahan baku berdasarkan ID.
      parameters:
      - description: ID Bahan
        in: path
        name: id
        required: true
        type: integer
      - description: Data update bahan utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BahanInput'
      produces:
      - application/json
      responses:
        "200":
          description: Bahan berhasil diupdate
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Bahan tidak ditemukan
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Master Bahan
      tags:
      - 05. Master Bahan Baku
  /api/bahan/reorder:
    put:
      consumes:
      - application/json
      description: Memperbarui posisi urutan bahan baku untuk tampilan di form.
      parameters:
      - description: 'Array of object key: id, urutan'
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/models.UrutanBahanInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Urutan bahan berhasil diperbarui
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update Urutan Tampilan Bahan
      tags:
      - 05. Master Bahan Baku
  /api/barangs:
    get:
      consumes:
      - application/json
      description: Mengambil daftar master barang lengkap dengan relasi resep dan
        kemasan, diurutkan berdasarkan urutan tampilan.
      produces:
      - application/json
      responses:
        "200":
          description: List data barang berhasil ditarik
          schema:
            items:
              $ref: '#/definitions/models.Barang'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil Seluruh Master Barang
      tags:
      - 03. Master Barang
    post:
      consumes:
      - application/json
      description: Menyimpan data master barang baru beserta kebutuhan kemasannya
        (kardus/plastik) jika ada.
      parameters:
      - description: Data Barang utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BarangInput'
      produces:
      - application/json
      responses:
        "200":
          description: Barang dan Kemasan berhasil dibuat
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON tidak valid
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Gagal menyimpan ke database
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat Master Barang Baru
      tags:
      - 03. Master Barang
  /api/barangs/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus master barang dari daftar aktif (masuk ke tong sampah).
      parameters:
      - description: ID Barang
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Barang berhasil dihapus
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus Master Barang (Soft Delete)
      tags:
      - 03. Master Barang
    put:
      consumes:
      - application/json
      description: Memperbarui data master barang beserta list kemasannya berdasarkan
        ID.
      parameters:
      - description: ID Barang
        in: path
        name: id
        required: true
        type: integer
      - description: Data Update Barang utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BarangInput'
      produces:
      - application/json
      responses:
        "200":
          description: Barang berhasil diupdate
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON tidak valid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Barang tidak ditemukan
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Master Barang
      tags:
      - 03. Master Barang
  /api/barangs/reorder:
    put:
      consumes:
      - application/json
      description: Memperbarui posisi/urutan barang untuk tampilan di form pembuatan
        nota. (Dipanggil saat Drag & Drop di UI Vue)
      parameters:
      - description: 'Array of object dengan key: id, urutan'
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/models.UrutanBarangInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Urutan barang berhasil diperbarui
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format data salah
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Urutan Tampilan Barang
      tags:
      - 03. Master Barang
  /api/catatan-besar:
    get:
      consumes:
      - application/json
      description: Mengambil laporan rekapitulasi kirim & retur barang yang difilter
        berdasarkan tanggal dan siklus.
      parameters:
      - description: 'Format YYYY-MM-DD (Default: Hari ini)'
        in: query
        name: tanggal
        type: string
      - description: Filter Siklus
        enum:
        - HARIAN
        - SiklusDua
        - SiklusKamisSenin
        - SiklusJumatSelasa
        - SiklusSabtuRabu
        in: query
        name: siklus
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menarik catatan besar
          schema:
            items:
              $ref: '#/definitions/models.CatatanBesarResponse'
            type: array
        "500":
          description: Kesalahan eksekusi database
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Laporan Catatan Besar
      tags:
      - 10. Laporan & Rangkuman
  /api/inventory/rusak:
    get:
      consumes:
      - application/json
      description: Menampilkan data barang terbuang/afkir harian.
      parameters:
      - description: Tanggal filter (YYYY-MM-DD)
        in: query
        name: tanggal
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil ditarik
          schema:
            items:
              $ref: '#/definitions/models.BarangRusak'
            type: array
      security:
      - BearerAuth: []
      summary: Ambil Riwayat Afkir / Gratisan
      tags:
      - 11. Operasional Inventory
    post:
      consumes:
      - application/json
      description: Menyimpan jumlah roti yang dibuang/gratis, akan otomatis mengurangi
        jumlah Sisa Layak Jual saat Tutup Buku nanti malam.
      parameters:
      - description: Data afkir
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BarangRusakInput'
      produces:
      - application/json
      responses:
        "200":
          description: Dicatat
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - BearerAuth: []
      summary: Catat Barang Rusak / Afkir
      tags:
      - 11. Operasional Inventory
  /api/inventory/rusak/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus riwayat afkir. Kalkulasi mesin Tutup Buku akan menyesuaikan
        sendiri.
      parameters:
      - description: ID Afkir
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dihapus
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - BearerAuth: []
      summary: Batal Catat Afkir
      tags:
      - 11. Operasional Inventory
  /api/kas:
    get:
      consumes:
      - application/json
      description: Menarik semua catatan transaksi masuk/keluar dari brankas secara
        menurun (terbaru di atas).
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menarik data kas
          schema:
            items:
              $ref: '#/definitions/models.TransaksiKas'
            type: array
        "500":
          description: Kesalahan eksekusi database
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil Riwayat Transaksi Kas
      tags:
      - 14. Kas & Keuangan
    post:
      consumes:
      - application/json
      description: Menyimpan transaksi kas secara manual, khusus untuk kategori non-sistem
        seperti pengeluaran rumah tangga atau setoran modal.
      parameters:
      - description: Data transaksi kas
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.KasInput'
      produces:
      - application/json
      responses:
        "200":
          description: Transaksi kas berhasil dicatat
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Format data JSON salah
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Kesalahan eksekusi database
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Catat Transaksi Kas Manual
      tags:
      - 14. Kas & Keuangan
  /api/kas/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus paksa catatan kas jika terjadi salah ketik atau salah
        input nominal.
      parameters:
      - description: ID Kas
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaksi kas berhasil dihapus
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus Transaksi Kas
      tags:
      - 14. Kas & Keuangan
  /api/konversi/sisa-kemarin:
    get:
      consumes:
      - application/json
      description: Mengambil data sisa layak jual hari-hari sebelumnya (berdasarkan
        batas kadaluwarsa barang) yang masih bisa diseret untuk pengiriman hari ini.
      parameters:
      - description: Tanggal pengiriman hari ini (YYYY-MM-DD)
        in: query
        name: tanggal
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data sisa kemarin ditarik
          schema:
            items:
              $ref: '#/definitions/models.SisaLayakJual'
            type: array
      security:
      - BearerAuth: []
      summary: Tarik Sisa Layak Jual (H-1)
      tags:
      - 11. Operasional Inventory
  /api/opname:
    get:
      consumes:
      - application/json
      description: Menampilkan histori koreksi paksa jumlah stok bahan fisik di gudang.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil ditarik
          schema:
            items:
              $ref: '#/definitions/models.StockOpname'
            type: array
      security:
      - BearerAuth: []
      summary: Ambil Riwayat Stock Opname (Sidak Fisik)
      tags:
      - 11. Operasional Inventory
    post:
      consumes:
      - application/json
      description: Memperbarui stok mutlak database dengan stok fisik nyata (menyesuaikan
        selisih/tumpah secara paksa).
      parameters:
      - description: Data penyesuaian stok fisik
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.StockOpnameInput'
      produces:
      - application/json
      responses:
        "200":
          description: Opname dicatat
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - BearerAuth: []
      summary: Eksekusi Stock Opname
      tags:
      - 11. Operasional Inventory
  /api/pembelian:
    get:
      consumes:
      - application/json
      description: Menarik riwayat belanja bahan fisik.
      parameters:
      - description: Tanggal Mulai (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: Tanggal Akhir (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil ditarik
          schema:
            items:
              $ref: '#/definitions/models.PembelianBahan'
            type: array
      security:
      - BearerAuth: []
      summary: Riwayat Pembelian Bahan Baku
      tags:
      - 11. Operasional Inventory
    post:
      consumes:
      - application/json
      description: Mencatat belanja bahan, otomatis menambah stok gudang, menimpa
        harga beli terbaru, dan memotong kas (jika lunas).
      parameters:
      - description: Data pembelian lengkap
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.PembelianBahanInput'
      produces:
      - application/json
      responses:
        "200":
          description: Pembelian berhasil dicatat
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Gagal potong kas/stok
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Catat Pembelian Bahan (Otomatis Tambah Stok)
      tags:
      - 11. Operasional Inventory
  /api/pembelian/{id}:
    delete:
      consumes:
      - application/json
      description: Membatalkan transaksi belanja, mengembalikan stok gudang ke semula,
        dan memulihkan saldo kas jika terlanjur lunas.
      parameters:
      - description: ID Pembelian
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dibatalkan
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - BearerAuth: []
      summary: Batalkan Pembelian (Rollback Stok & Kas)
      tags:
      - 11. Operasional Inventory
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: "Bearer {token_jwt}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"